Introducing FhirEncounterService
