+++ src/RestControllers/FhirEncounterRestController.php
@@ -12,19 +12,19 @@
 
 namespace OpenEMR\RestControllers;
 
-use OpenEMR\Services\EncounterService;
+use OpenEMR\Services\FhirEncounterService;
 use OpenEMR\Services\FhirResourcesService;
 use OpenEMR\RestControllers\RestControllerHelper;
 use OpenEMR\FHIR\R4\FHIRResource\FHIRBundle\FHIRBundleEntry;
 
 class FhirEncounterRestController
 {
-    private $encounterService;
+    private $fhirEncounterService;
     private $fhirService;
 
     public function __construct()
     {
-        $this->encounterService = new EncounterService();
+        $this->fhirEncounterService = new FhirEncounterService();
         $this->fhirService = new FhirResourcesService();
     }
 
@@ -32,8 +32,8 @@ class FhirEncounterRestController
 
     public function getOne($eid)
     {
-        $oept = $this->encounterService->getEncounter($eid);
-        $encounterResource = $this->fhirService->createEncounterResource($eid, $oept, false);
+        $oept = $this->fhirEncounterService->getEncounter($eid);
+        $encounterResource = $this->fhirEncounterService->createEncounterResource($eid, $oept, false);
 
         return RestControllerHelper::responseHandler($encounterResource, null, 200);
     }
@@ -49,14 +49,14 @@ class FhirEncounterRestController
             'pid' => $search['patient'],
             'provider_id' => $search['practitioner']);
 
-        $searchResult = $this->encounterService->getEncountersBySearch($searchParam);
+        $searchResult = $this->fhirEncounterService->getEncountersBySearch($searchParam);
         if ($searchResult === false) {
             http_response_code(404);
             exit;
         }
         $entries = array();
         foreach ($searchResult as $oept) {
-            $entryResource = $this->fhirService->createEncounterResource($oept['encounter'], $oept, false);
+            $entryResource = $this->fhirEncounterService->createEncounterResource($oept['encounter'], $oept, false);
             $entry = array(
                 'fullUrl' => $resourceURL . "/" . $oept['encounter'],
                 'resource' => $entryResource
