+++ src/Services/FhirResourcesService.php
@@ -12,26 +12,16 @@
 
 namespace OpenEMR\Services;
 
-use OpenEMR\FHIR\R4\FHIRDomainResource\FHIREncounter;
-use OpenEMR\FHIR\R4\FHIRDomainResource\FHIRPatient;
 use OpenEMR\FHIR\R4\FHIRDomainResource\FHIRPractitioner;
 use OpenEMR\FHIR\R4\FHIRElement\FHIRAddress;
 use OpenEMR\FHIR\R4\FHIRElement\FHIRAdministrativeGender;
 use OpenEMR\FHIR\R4\FHIRElement\FHIRHumanName;
 use OpenEMR\FHIR\R4\FHIRElement\FHIRId;
-use OpenEMR\FHIR\R4\FHIRElement\FHIRReference;
-use OpenEMR\FHIR\R4\FHIRResource\FHIREncounter\FHIREncounterParticipant;
 use OpenEMR\FHIR\R4\FHIRResource\FHIRBundle;
 use OpenEMR\FHIR\R4\FHIRResource\FHIRBundle\FHIRBundleLink;
 use OpenEMR\FHIR\R4\PHPFHIRResponseParser;
 
 use OpenEMR\Services\ProviderService;
-use OpenEMR\Services\FhirValidationService;
-
-//use OpenEMR\FHIR\R4\FHIRResource\FHIREncounter\FHIREncounterLocation;
-//use OpenEMR\FHIR\R4\FHIRResource\FHIREncounter\FHIREncounterDiagnosis;
-//use OpenEMR\FHIR\R4\FHIRElement\FHIRPeriod;
-//use OpenEMR\FHIR\R4\FHIRElement\FHIRParticipantRequired;
 
 class FhirResourcesService
 {
@@ -87,37 +77,6 @@ class FhirResourcesService
         $resource->setGender($gender);
         $resource->addName($name);
         $resource->addAddress($address);
-
-        if ($encode) {
-            return json_encode($resource);
-        } else {
-            return $resource;
-        }
-    }
-
-    public function createEncounterResource($eid = '', $data = '', $encode = true)
-    {
-        $pid = $data['pid'];
-        //$temp = $data['provider_id'];
-        //$r = $this->createPractitionerResource($data['provider_id'], $temp);
-        $resource = new FHIREncounter();
-        $id = new FhirId();
-        $id->setValue($eid);
-        $resource->setId($id);
-        $participant = new FHIREncounterParticipant();
-        $prtref = new FHIRReference;
-        $temp = 'Practitioner/' . $data['provider_id'];
-        $prtref->setReference($temp);
-        $participant->setIndividual($prtref);
-        $date = date('Y-m-d', strtotime($data['date']));
-        $participant->setPeriod(['start' => $date]);
-
-        $resource->addParticipant($participant);
-        $reason = new FHIRCodeableConcept();
-        $reason->setText($data['reason']);
-        $resource->addReasonCode($reason);
-        $resource->status = 'finished';
-        $resource->setSubject(['reference' => "Patient/$pid"]);
 
         if ($encode) {
             return json_encode($resource);
